feat: modularize evaluation
